+++ src/cmd_parser.cpp
@@ -144,7 +144,6 @@ bool CmdParser::parseNextCommand()
       std::vector<Expr> params;
       // attributes marked on variables
       std::map<ExprValue*, AttrMap> pattrMap;
-      bool flattenFunction = (tok != Token::DECLARE_ORACLE_FUN);
       if (tok == Token::DECLARE_FUN || tok == Token::DECLARE_ORACLE_FUN)
       {
         sorts = d_eparser.parseTypeList();
@@ -162,6 +161,10 @@ bool CmdParser::parseNextCommand()
       sk = Kind::CONST;
       if (tok==Token::DECLARE_ORACLE_FUN)
       {
+        if (sorts.empty())
+        {
+          d_lex.parseError("Oracle functions must have at least one argument");
+        }
         ck = Attr::ORACLE;
         sk = Kind::ORACLE;
         std::string oname = d_eparser.parseSymbol();
@@ -180,7 +183,9 @@ bool CmdParser::parseNextCommand()
       t = ret;
       if (!sorts.empty())
       {
-        t = d_state.mkFunctionType(sorts, ret, flattenFunction);
+        t = (tok == Token::DECLARE_ORACLE_FUN)
+                ? d_state.mkProgramType(sorts, ret)
+                : d_state.mkFunctionType(sorts, ret);
       }
       std::vector<Expr> opaqueArgs;
       while (t.getKind()==Kind::FUNCTION_TYPE && t[0].getKind()==Kind::OPAQUE_TYPE)
@@ -359,14 +364,6 @@ bool CmdParser::parseNextCommand()
       }
       d_state.pushScope();
       std::string name = d_eparser.parseSymbol();
-      if (d_lex.peekToken()==Token::KEYWORD)
-      {
-        std::string keyword = d_eparser.parseKeyword();
-        if (keyword!="ethos")
-        {
-          d_lex.parseError("Unsupported rule format");
-        }
-      }
       std::vector<Expr> vs =
           d_eparser.parseAndBindSortedVarList(Kind::PROOF_RULE);
       Expr assume;
@@ -716,14 +713,6 @@ bool CmdParser::parseNextCommand()
     case Token::PROGRAM:
     {
       std::string name = d_eparser.parseSymbol();
-      if (d_lex.peekToken()==Token::KEYWORD)
-      {
-        std::string keyword = d_eparser.parseKeyword();
-        if (keyword!="ethos")
-        {
-          d_lex.parseError("Unsupported program format");
-        }
-      }
       // push the scope
       d_state.pushScope();
       std::vector<Expr> vars =
@@ -733,7 +722,11 @@ bool CmdParser::parseNextCommand()
       Expr progType = retType;
       if (!argTypes.empty())
       {
-        progType = d_state.mkFunctionType(argTypes, retType, false);
+        progType = d_state.mkProgramType(argTypes, retType);
+      }
+      else
+      {
+        d_lex.parseError("Programs must have at least one argument");
       }
       // it may have been forward declared
       Expr pprev = d_state.getVar(name);