+++ src/type_checker.cpp
@@ -240,7 +240,8 @@ Expr TypeChecker::getTypeInternal(ExprValue* e, std::ostream* out)
     case Kind::TYPE:
     case Kind::ABSTRACT_TYPE:
     case Kind::BOOL_TYPE:
-    case Kind::FUNCTION_TYPE: return d_state.mkType();
+    case Kind::FUNCTION_TYPE:
+    case Kind::PROGRAM_TYPE: return d_state.mkType();
     case Kind::PROOF_TYPE:
     {
       ExprValue* ctype = d_state.lookupType(e->d_children[0]);
@@ -342,7 +343,8 @@ Expr TypeChecker::getTypeAppInternal(std::vector<ExprValue*>& children,
   ExprValue* hd = children[0];
   ExprValue* hdType = d_state.lookupType(hd);
   Assert(hdType != nullptr) << "No type for " << Expr(hd);
-  if (hdType->getKind() != Kind::FUNCTION_TYPE)
+  Kind hk = hdType->getKind();
+  if (hk != Kind::FUNCTION_TYPE && hk != Kind::PROGRAM_TYPE)
   {
     // non-function at head
     if (out)
@@ -1597,7 +1599,6 @@ bool TypeChecker::computedParameterizedInternal(AppInfo* ai,
     {
       // if not in an application, we fail
       Warning() << "Failed to determine parameters for " << hd << std::endl;
-      AlwaysAssert(false);
       return false;
     }
     else